dm lossy: streamline files
Merge common.[hc] into the dm-lossy-target.c and
update makefiles to use the new target name.

Also remove perforce tags and other unnecessary bits
from files.
