+++ src/c++/devices/lossy/dkms.conf
@@ -4,15 +4,14 @@
 #
 # %LICENSE%
 #
-# DKMS configuration for the Fua test module
+# DKMS configuration for the lossy module
 #
-# $Id$
 
-PACKAGE_NAME="pbitdory"
+PACKAGE_NAME="pbitlossy"
 PACKAGE_VERSION="1"
 MODULE_NAME="${PACKAGE_NAME}"
 AUTOINSTALL="yes"
 
-BUILT_MODULE_NAME[0]="pbitdory"
+BUILT_MODULE_NAME[0]="pbitlossy"
 DEST_MODULE_LOCATION[0]="/kernel/drivers/block/"
 STRIP[0]="no"
