+++ src/c++/devices/lossy/dm-lossy-target.h
@@ -2,8 +2,6 @@
  * %COPYRIGHT%
  *
  * %LICENSE%
- *
- * $Id$
  */
 
 #include <linux/module.h>
